% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installDockerImage.R
\name{installDockerImage}
\alias{installDockerImage}
\title{Install timsconvert docker image}
\usage{
installDockerImage(imageURL = "ghcr.io/thomas-enzlein/timsconvert:main")
}
\arguments{
\item{imageURL}{character, the URL of the image.}
}
\description{
Install timsconvert docker image
}
